move pages down a level for more resilient routing
